refactor database & add validation

+++ src/main/java/com/worldbestsoft/model/Catalog.java
@@ -1,6 +1,6 @@
 package com.worldbestsoft.model;
 
-// Generated Jul 16, 2013 3:46:49 PM by Hibernate Tools 4.0.0
+// Generated Jul 23, 2013 9:54:03 PM by Hibernate Tools 4.0.0
 
 import java.math.BigDecimal;
 import java.util.Date;
@@ -26,7 +26,7 @@ import javax.persistence.TemporalType;
 @Table(name = "catalog", catalog = "onion")
 public class Catalog implements java.io.Serializable {
 
-	private Integer id;
+	private Long id;
 	private InvItem invItem;
 	private CatalogType catalogType;
 	private String code;
@@ -68,11 +68,11 @@ public class Catalog implements java.io.Serializable {
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
 	@Column(name = "id", unique = true, nullable = false)
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
