+++ src/main/java/com/worldbestsoft/model/CatalogItem.java
@@ -1,7 +1,8 @@
 package com.worldbestsoft.model;
 
-// Generated Jul 16, 2013 3:46:49 PM by Hibernate Tools 4.0.0
+// Generated Jul 23, 2013 9:54:03 PM by Hibernate Tools 4.0.0
 
+import java.math.BigDecimal;
 import java.util.Date;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -22,11 +23,11 @@ import javax.persistence.TemporalType;
 @Table(name = "catalog_item", catalog = "onion")
 public class CatalogItem implements java.io.Serializable {
 
-	private Integer id;
+	private Long id;
 	private Catalog catalog;
 	private InvItem invItem;
 	private String name;
-	private Integer qty;
+	private BigDecimal qty;
 	private Date createDate;
 	private String createUser;
 	private Date updateDate;
@@ -40,7 +41,7 @@ public class CatalogItem implements java.io.Serializable {
 		this.invItem = invItem;
 	}
 
-	public CatalogItem(Catalog catalog, InvItem invItem, String name, Integer qty, Date createDate, String createUser, Date updateDate, String updateUser) {
+	public CatalogItem(Catalog catalog, InvItem invItem, String name, BigDecimal qty, Date createDate, String createUser, Date updateDate, String updateUser) {
 		this.catalog = catalog;
 		this.invItem = invItem;
 		this.name = name;
@@ -54,11 +55,11 @@ public class CatalogItem implements java.io.Serializable {
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
 	@Column(name = "id", unique = true, nullable = false)
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -91,12 +92,12 @@ public class CatalogItem implements java.io.Serializable {
 		this.name = name;
 	}
 
-	@Column(name = "qty")
-	public Integer getQty() {
+	@Column(name = "qty", precision = 10)
+	public BigDecimal getQty() {
 		return this.qty;
 	}
 
-	public void setQty(Integer qty) {
+	public void setQty(BigDecimal qty) {
 		this.qty = qty;
 	}
